Beta 1.0.1

+++ lib/feature/data/models/attendance_stat_model.dart
@@ -4,7 +4,7 @@ class AttendanceStatModel {
   final int totalClasses;
   final int presentClasses;
   final int absentClasses;
-  final int attendancePercentage;
+  final double attendancePercentage;
 
   AttendanceStatModel({
     required this.studentId,
